- Split components.h into multiple files
